Refactored GKE code support
