+++ edbdeploy/project.py
@@ -70,8 +70,9 @@ class Project:
  
     terraform_templates = ['variables.tf.template', 'tags.tf.template']
     ansible_collection_name = 'edb_devops.edb_postgres:>=%s,<4.0.0' % __edb_ansible_version__  # noqa
-    k8s_core_ansible_collection_name = 'kubernetes.core:>=%s,<2.3.0' % __edb_ansible_version__  # noqa
-    k8s_community_ansible_collection_name = 'kubernetes.core:>=%s,<2.0.1' % __edb_ansible_version__  # noqa        
+    # No need to define version for kubernetes core
+    k8s_core_ansible_collection_name = 'kubernetes.core'  # noqa    
+    k8s_community_ansible_collection_name = 'kubernetes.core' # noqa
 
     def __init__(self, cloud, name, env, bin_path=None):
         self.env = env
@@ -1755,153 +1756,3 @@ class Project:
             self.ansible_vars['route53_session_token'] = n_route53_session_token
             self._save_ansible_vars()
             
-    """
-    Kubernetes related methods
-    """
-    def gcloud_gke_configure(self, env):
-        """
-        Configure sub-comand for Google Cloud GKE environment
-        """
-        # Load specifications
-        env.cloud_spec = self._load_cloud_specs(env)
-                
-        # Copy the kubernetes role in ansible project directory
-        ansible_roles_path = os.path.join(self.project_path, "roles")
-
-        # Hook function called by Project.configure()
-        # Transform Terraform templates
-        self._transform_terraform_tpl()
-        # Build the vars files for Terraform and Ansible
-        self._build_terraform_vars_file(env)
-        # Load terraform variables
-        self._load_terraform_vars()
-        # Load cnpType
-        cnpType = self.terraform_vars['cnpType']        
-        # Assign ansible playbook to copy
-        match cnpType:
-            case 'pg':
-                self.ansible_vars = { 'reference_architecture': 'setup_cloudnativepg_sandbox', }
-            case 'postgres':
-                self.ansible_vars = { 'reference_architecture': 'setup_cloudnativepostgres_sandbox', }
-            case _:
-                self.ansible_vars = { 'reference_architecture': 'setup_cloudnativepg_sandbox', }
-        # Copy Ansible playbook into project dir.
-        self._copy_ansible_playbook()
-        
-        # Initialize Terraform so that a project removal 
-        # can immediately be followed successfully
-        # Load terraform variables
-        self._load_terraform_vars()
-        
-        terraform = TerraformCli(
-            self.project_path, self.terraform_plugin_cache_path,
-            bin_path=self.cloud_tools_bin_path
-        )
-        
-        with AM("Google Cloud GKE Terraform project initialization"):
-            self.update_state('terraform', 'INITIALIZATING')
-            terraform.init()
-            self.update_state('terraform', 'INITIALIZATED')
-
-        # note: we do not raise any AnsibleCliError from this function
-        # because AnsibleCliError are used to trigger stuffs when they are
-        # catched. In this case, we do not want trigger anything if something
-        # fails.
-        try:
-            output = exec_shell(
-                [self.bin("ansible-playbook"), "--version"],
-                environ=self.environ
-            )
-        except CalledProcessError as e:
-            logging.error("Failed to execute the command: %s", e.cmd)
-            logging.error("Return code is: %s", e.returncode)
-            logging.error("Output: %s", e.output)
-            raise Exception(
-                "Terraform executable seems to be missing. Please install it "
-                "or check your PATH variable"
-            )            
-
-    def gcloud_gke_provision(self, env):
-        # Load terraform variables
-        self._load_terraform_vars()
-        
-        terraform = TerraformCli(
-            self.project_path, self.terraform_plugin_cache_path,
-            bin_path=self.cloud_tools_bin_path
-        )
-
-        with AM("Google Cloud GKE Terraform project initialization"):
-            self.update_state('terraform', 'INITIALIZATING')
-            terraform.init()
-            self.update_state('terraform', 'INITIALIZATED')
-
-        with AM("Google Cloud GKE Terraform project provisioning"):
-            self.update_state('terraform', 'PROVISIONING')
-            terraform.apply(self.terraform_vars_file)
-            self.update_state('terraform', 'PROVISIONED')
-
-        # Assign region from Terraform variables
-        region = self.terraform_vars['gcpRegion']
-        kClusterName = self.terraform_vars['kClusterName']        
-        # Get Kubernetes Cluster credentials
-        output = exec_shell(
-            [self.bin("gcloud"), "container", "clusters", "get-credentials", 
-                kClusterName, "--region", region],
-            environ=self.environ
-        )            
-            
-    def gcloud_gke_deploy(self, no_install_collection, pre_deploy_ansible=None,
-                   post_deploy_ansible=None, skip_main_playbook=False,
-                   disable_pipelining=False):
-        """
-        Deployment method for the Google Cloud GKE environments
-        """
-        # Load terraform variables
-        self._load_terraform_vars()
-       
-        # Assign region from Terraform variables
-        region = self.terraform_vars['gcpRegion']
-        kClusterName = self.terraform_vars['kClusterName']        
-        # Get Kubernetes Cluster credentials
-        output = exec_shell(
-            [self.bin("gcloud"), "container", "clusters", "get-credentials", 
-                kClusterName, "--region", region],
-            environ=self.environ
-        )
-
-        inventory_data = None
-        ansible = AnsibleCli(
-            self.project_path,
-            bin_path=self.cloud_tools_bin_path
-        )
-
-        if not skip_main_playbook:
-            self.update_state('ansible', 'DEPLOYING')
-            with AM("Deploying components with Ansible"):
-                ansible.run_playbook_minimal(
-                    self.cloud,
-                    self.ansible_playbook,
-                    disable_pipelining=disable_pipelining,
-                )
-            self.update_state('ansible', 'DEPLOYED')
-
-    def gcloud_gke_destroy(self):
-        """
-        GCloud GKE destroy method
-        """
-        terraform = TerraformCli(
-            self.project_path, self.terraform_plugin_cache_path,
-            bin_path=self.cloud_tools_bin_path
-        )
-
-        inventory_data = None
-        ansible = AnsibleCli(
-            self.project_path,
-            bin_path=self.cloud_tools_bin_path
-        )
-
-        with AM("Destroying cloud resources"):
-            self.update_state('terraform', 'DESTROYING')
-            terraform.destroy(self.terraform_vars_file)
-            self.update_state('terraform', 'DESTROYED')
-            self.update_state('ansible', 'UNKNOWN')
